feat: add new service class
